refactor: split semantic tokens
