+++ OUDS/Core/Tokens/SemanticTokens/Sources/Aliases/BorderSemanticTokens+Aliases.swift
@@ -1,0 +1,24 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import Foundation
+import OUDSTokensRaw
+
+/// Type alias precising `BorderWidthRawToken` is used as a value of this **border width semantic token**, to keep grammar clean and clear with design system grammar.
+public typealias BorderWidthSemanticToken = BorderWidthRawToken
+
+/// Type alias precising `BorderRadiusRawToken` is used as a value of this **border raw semantic token**, to keep grammar clean and clear with design system grammar.
+public typealias BorderRadiusSemanticToken = BorderRadiusRawToken
+
+/// Type alias precising `BorderStyleRawToken` is used as a value of this **border style semantic token**, to keep grammar clean and clear with design system grammar.
+public typealias BorderStyleSemanticToken = BorderStyleRawToken
