+++ OUDS/Core/Tokens/SemanticTokens/Sources/Aliases/ElevationSemanticTokens+Aliases.swift
@@ -1,0 +1,33 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+// 
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+// 
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System 
+//
+
+import Foundation
+import OUDSTokensRaw
+
+/// Basically an elevation semantic token for Z Index is a raw token for elevation, with the same final type, to keep grammar clean and clear with design system grammar.
+public typealias ElevationZIndexSemanticToken = ElevationRawToken
+
+/// Basically an elevation semantic token for X offset is a raw token for elevation, with the same final type, to keep grammar clean and clear with design system grammar.
+public typealias ElevationXSemanticToken = ElevationRawToken
+
+/// Basically an elevation semantic token for Y Index is a raw token for elevation, with the same final type, to keep grammar clean and clear with design system grammar.
+public typealias ElevationYSemanticToken = ElevationRawToken
+
+/// Basically an elevation semantic token for blur effect is a raw token for elevation, with the same final type, to keep grammar clean and clear with design system grammar.
+public typealias ElevationBlurSemanticToken = ElevationRawToken
+
+/// Basically an elevation semantic token for shadow colors is a raw token for colors, to keep grammar clean and clear with design system grammar.
+public typealias ElevationColorSemanticToken = ColorRawToken
+
+/// Basically an elevation semantic token for box shadow is a raw token for box shadow, with the same final type, to keep grammar clean and clear with design system grammar.
+public typealias ElevationCompositeSemanticToken = ElevationCompositeRawToken
