+++ OUDS/Core/Tokens/SemanticTokens/Sources/Aliases/GridSemanticTokens+Aliases.swift
@@ -1,0 +1,24 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+// 
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+// 
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System 
+//
+
+import Foundation
+import OUDSTokensRaw
+
+/// Basically a grid semantic token for iOS Extra Compact values is a grid raw token, to keep grammar clean and clear with design system grammar.
+public typealias GridExtraCompactSemanticToken = GridRawToken
+
+/// Basically a grid semantic token for iOS Compact values is a grid raw token, to keep grammar clean and clear with design system grammar.
+public typealias GridCompactSemanticToken = GridRawToken
+
+/// Basically a grid semantic token for iOS Regulard values is a grid raw token, to keep grammar clean and clear with design system grammar.
+public typealias GridRegularSemanticToken = GridRawToken
