+++ OUDS/Core/Tokens/SemanticTokens/Sources/Aliases/OpacitySemanticTokens+Aliases.swift
@@ -1,0 +1,18 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+// 
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+// 
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System 
+//
+
+import Foundation
+import OUDSTokensRaw
+
+/// Basically an opacity semantic token is an opacity raw token, to keep grammar clean and clear with design system grammar.
+public typealias OpacitySemanticToken = OpacityRawToken
