+++ OUDS/Core/Tokens/SemanticTokens/Sources/Aliases/SizingSemanticTokens+Aliases.swift
@@ -1,0 +1,21 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import Foundation
+import OUDSTokensRaw
+
+/// Basically a size semantic token for width and height is a dimension raw token, it has the same final type, to keep grammar clean and clear with design system grammar.
+public typealias SizingWidthHeightSemanticToken = DimensionRawToken
+
+/// Basically a size semantic token for max width is a dimension raw token, it has the same final type, to keep grammar clean and clear with design system grammar.
+public typealias SizingMaxWidthSemanticToken = DimensionRawToken
