+++ OUDS/Core/Tokens/SemanticTokens/Sources/Aliases/SpacingSemanticTokens+Aliases.swift
@@ -1,0 +1,33 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import Foundation
+import OUDSTokensRaw
+
+/// Basically a space semantic token for layout is a dimension raw token, it has the same final type, to keep grammar clean and clear with design system grammar.
+public typealias SpacingLayoutSemanticToken = DimensionRawToken
+
+/// Basically a space semantic token for padding inline is a dimension raw token, it has the same final type, to keep grammar clean and clear with design system grammar.
+public typealias SpacingPaddingInlineSemanticToken = DimensionRawToken
+
+/// Basically a space semantic token for padding block is a dimension raw token, it has the same final type, to keep grammar clean and clear with design system grammar.
+public typealias SpacingPaddingBlockSemanticToken = DimensionRawToken
+
+/// Basically a space semantic token for padding inset is a dimension raw token, it has the same final type, to keep grammar clean and clear with design system grammar.
+public typealias SpacingPaddingInsetSemanticToken = DimensionRawToken
+
+/// Basically a space semantic token for gap inline is a dimension raw token, it has the same final type, to keep grammar clean and clear with design system grammar.
+public typealias SpacingGapInlineSemanticToken = DimensionRawToken
+
+/// Basically a space semantic token for gap stack is a dimension raw token, it has the same final type, to keep grammar clean and clear with design system grammar.
+public typealias SpacingGapStackSemanticToken = DimensionRawToken
