+++ OUDS/Core/Tokens/SemanticTokens/Sources/Aliases/TypographySemanticTokens+Aliases.swift
@@ -1,0 +1,27 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import Foundation
+import OUDSTokensRaw
+
+/// The global design system tools uses verbs of semantic token for font family, which is basically a raw token for font family
+public typealias TypographyFontFamilySemanticToken = TypographyFontFamilyRawToken
+
+/// The global design system tools uses verbs of semantic token for font weight, which is basically a raw token for font weight
+public typealias TypographyFontWeightSemanticToken = TypographyFontWeightRawToken
+
+/// The global design system tools uses verbs of semantic token for font size, which is basically a raw token for font size
+public typealias TypographyFontSizeSemanticToken = TypographyFontSizeRawToken
+
+/// The global design system tools uses verbs of semantic token for font line height, which is basically a raw token for font line height
+public typealias TypographyFontLineHeightSemanticToken = TypographyFontLineHeightRawToken
