+++ OUDS/Core/Tokens/SemanticTokens/Sources/Values/SizingSemanticTokens.swift
@@ -14,39 +14,6 @@
 import Foundation
 import OUDSTokensRaw
 
-// MARK: - Type aliases to keep grammar clear
-
-/// Basically a size semantic token for width and height is a dimension raw token, it has the same final type
-public typealias SizingWidthHeightSemanticToken = DimensionRawToken
-
-/// Basically a size semantic token for max width is a dimension raw token, it has the same final type
-public typealias SizingMaxWidthSemanticToken = DimensionRawToken
-
-// MARK: Semantic tokens
-
-// MARK: - Composite Semantic Token
-
-/// Composite semantic tokens which will wrap a combination of `SizingWidthHeightSemanticToken` depending to viewports.
-public final class SizingCompositeSemanticToken: NSObject {
-
-    /// For **extra-compact** and **compact** viewports
-    public let compact: SizingWidthHeightSemanticToken
-    /// For **regular** and **medium** viewports
-    public let regular: SizingWidthHeightSemanticToken
-
-    public init(compact: SizingWidthHeightSemanticToken, regular: SizingWidthHeightSemanticToken) {
-        self.compact = compact
-        self.regular = regular
-    }
-
-    public override func isEqual(_ object: Any?) -> Bool {
-        guard let other = object as? SizingCompositeSemanticToken else { return false }
-        return self.compact == other.compact && self.regular == other.regular
-    }
-}
-
-// MARK: - Sizing Semantic Tokens
-
 /// This is a group of semantic tokens for **sizing**.
 /// It defines all `SizingSemanticToken` a theme must have.
 public protocol SizingSemanticTokens {
