+++ OUDS/Core/Tokens/SemanticTokens/Sources/Values/SpacingSemanticTokens.swift
@@ -14,51 +14,6 @@
 import Foundation
 import OUDSTokensRaw
 
-// MARK: - Type aliases to keep grammar clear
-
-/// Basically a space semantic token for layout is a dimension raw token, it has the same final type
-public typealias SpacingLayoutSemanticToken = DimensionRawToken
-
-/// Basically a space semantic token for padding inline is a dimension raw token, it has the same final type
-public typealias SpacingPaddingInlineSemanticToken = DimensionRawToken
-
-/// Basically a space semantic token for padding block is a dimension raw token, it has the same final type
-public typealias SpacingPaddingBlockSemanticToken = DimensionRawToken
-
-/// Basically a space semantic token for padding inset is a dimension raw token, it has the same final type
-public typealias SpacingPaddingInsetSemanticToken = DimensionRawToken
-
-/// Basically a space semantic token for gap inline is a dimension raw token, it has the same final type
-public typealias SpacingGapInlineSemanticToken = DimensionRawToken
-
-/// Basically a space semantic token for gap stack is a dimension raw token, it has the same final type
-public typealias SpacingGapStackSemanticToken = DimensionRawToken
-
-// MARK: Semantic tokens
-
-// MARK: - Composite Semantic Token
-
-/// Composite semantic tokens which will wrap a combination of `DimensionRawToken` depending to viewports.
-public final class SpacingCompositeSemanticToken: NSObject {
-
-    /// For **extra-compact** and **compact** viewports
-    public let compact: DimensionRawToken
-    /// For **regular** and **medium** viewports
-    public let regular: DimensionRawToken
-
-    public init(compact: DimensionRawToken, regular: DimensionRawToken) {
-        self.compact = compact
-        self.regular = regular
-    }
-
-    public override func isEqual(_ object: Any?) -> Bool {
-        guard let other = object as? SpacingCompositeSemanticToken else { return false }
-        return self.compact == other.compact && self.regular == other.regular
-    }
-}
-
-// MARK: - Spacing Semantic Tokens
-
 /// This is a group of semantic tokens for **spacing**.
 /// It defines all `SpacingSemanticToken` a theme must have.
 public protocol SpacingSemanticTokens {
